separate extensions / plugins tests
