+++ test/plugins/genbank.spec.js
@@ -98,7 +98,7 @@ describe('Plugins', () => {
 
     //todo - this should use a more correct example file, and should validate the blocks
     //fixme - this test looks like it should fail? what is 'genbank' into exportBlock()?
-    it('should export block to Genbank', function exportGB(done) {
+    it.skip('should export block to Genbank', function exportGB(done) {
       fs.readFile(path.resolve(__dirname, '../res/sampleBlocks.json'), 'utf8', (err, sampleBlocksJson) => {
         const sampleBlocks = JSON.parse(sampleBlocksJson);
         exportBlock('genbank', sampleBlocks)
